fix(codec): decode a "!!int" to the Go type an untagged integer takes
castToInteger narrowed a tagged integer to int, while the same number
untagged decodes to uint64, or int64 when negative. "!!int 1" and "1" were
then two keys to a Go map, so an own "!!int 1" did not beat the "1" a "<<"
brought in, and the mapping held both entries.

A tagged integer now keeps the type taggedInteger reads it at -- uint64,
int64 or *big.Int -- as the untagged spelling does, and as an empty "!!int"
already did. A caller asserting int on "k: !!int 1" now gets uint64.

+++ codec/zz_lax_tags_test.go
@@ -63,7 +63,7 @@ func TestLaxTagsReadsTheTextTheScalarWasWrittenWith(t *testing.T) {
 func TestLaxTagsLeavesEveryOtherAnswerAlone(t *testing.T) {
 	for src, want := range map[string]any{
 		// A tag naming what the scalar is.
-		"k: !!int 5\n":     5,
+		"k: !!int 5\n":     uint64(5),
 		"k: !!bool true\n": true,
 		"k: !!str 0x10\n":  "0x10",
 		// A tag this library has no rule for is read by its kind, option or no.
